+++ Libs/YahnisElsts/PluginUpdateChecker/Puc/v5p2/UpgraderStatus.php
@@ -1,5 +1,5 @@
 <?php
-namespace WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2;
+namespace WordPress\Ppfeufer\Plugin\WpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2;
 
 if ( !class_exists(UpgraderStatus::class, false) ):