+++ MemoryUsage.php
@@ -11,13 +11,13 @@
  * Domain Path: /l10n
  */
 
-namespace WordPress\Plugins\PpWpMemoryUsage;
+namespace WordPress\Ppfeufer\Plugin\WpMemoryUsage;
 
 require_once(
     trailingslashit(value: __DIR__) . 'Libs/YahnisElsts/PluginUpdateChecker/plugin-update-checker.php'
 );
 
-use WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5\PucFactory;
+use WordPress\Ppfeufer\Plugin\WpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5\PucFactory;
 
 class MemoryUsage {
     /**